+++ api/src/models/units/CodeKataUnit.ts
@@ -124,7 +124,7 @@ codeKataSchema.methods.toHtmlForIndividualPDF = function (): String {
   html += '<div>' +  md.render('<div class="codeBox">' + md.render(this.test) + '</div>') + '</div>';
   html += '</div>';
 
-  html += '</div ><div><h2>Solutions</h2></div>';
+  html += '</div ><div><h2>Solution</h2></div>';
   html += '<h5>Task</h5>';
   html += '<div>' +  md.render('<div class="codeBox">' + md.render(this.definition) + '</div>') + '</div>';
   html += '<h5>Code</h5>';
@@ -139,7 +139,7 @@ codeKataSchema.methods.toHtmlForSinglePDF = function (): String {
 
 
   let html = '<div class="bottomBoxWrapper">';
-  html += '<div><h4>' + md.render('Unit:' + this.name) + '</h4>'
+  html += '<div><h4>' + md.render('Unit: ' + this.name) + '</h4>'
     + '<span>' + md.render(this.description ? 'Description: ' + this.description : '') + '</span></div>';
   html += '<h5>Task</h5>';
   html += '<div>' +  md.render('<div class="codeBox">' + md.render(this.definition) + '</div>') + '</div>';
@@ -154,7 +154,7 @@ codeKataSchema.methods.toHtmlForSinglePDFSolutions = function (): String {
   const md = new MarkdownIt({html: true});
 
   let html = '';
-  html += '<div><h4>' + md.render('Unit:' + this.name) + '</h4></div>';
+  html += '<div><h4>' + md.render('Unit: ' + this.name) + '</h4></div>';
   html += '<h5>Task</h5>';
   html += '<div>' +  md.render('<div class="codeBox">' + md.render(this.definition) + '</div>') + '</div>';
   html += '<h5>Code</h5>';